Re #5460 Removed duplicate fields in minAmountXXX fields Customer, CA, BA, UA. Changed cascase to REMOVE from ALL for dependent entities for GDRP. Changed default code format to support spaces. Reverted back FieldAuditInterceptor.findDirty() method
super.findDirty() returns null - which according to that method javadoc means that its up to hibernate to do dirty checking.
AND then you check if CFs are dirty and add cfValue to the list of dirty parameters if applicable.
So as result, if any CF value was changed, all that it would return is ["cfValues"] as dirty fieldnames. That means that all other field changes are ignored.