Re #5489 Fixed rerating when charge period is rerated a second time and only a portion of an first charge is rerated again.
Example: initial charge WO was 01/10 to 01/11. A change came in on 22/10 and two WOs were created instead: 01/10-22/10 and 22/10-01/11.
A second change came in on 27/10 and three WOs were created instead: 01/10-22/10, 22/10-27/10 and 27/10-01/11.
When processing a second change and rerating a period 22/10-01/11, it should be aware that proration shall be applied as initial full period was 01/10-01/11